+++ PaperMalKing.AniList.Wrapper/Models/Interfaces/IIdentifiable.cs
@@ -4,5 +4,5 @@ namespace PaperMalKing.AniList.Wrapper.Models.Interfaces;
 
 internal interface IIdentifiable
 {
-	ulong Id { get; init; }
+	uint Id { get; init; }
 }
